requirements.txt
simplematrixbotlib==2.12.0
python-dotenv
pyyaml
apscheduler
pytz
